plagiat2/nearest.py
from abc import abstractmethod, ABC
from collections import OrderedDict
import numpy as np
  

import torch
from ..config import prepare_config
     
from sklearn.metrics import average_precision_score, roc_curve
from .knn import KNNIndex

class ATRBase(nearestneighboursbase):
   

    @property
  
    def match_self(self):
    
   
        return self._match_self
  

    @abstractmethod
    def _aggreg_ate(self, nearest_same, nearest_scores, num, class_si, positive_scores, confidences=None):
        pass

    def __init__(self, match_self=False):
        super().__init__()
        self._match_self = match_self

    @property
    def NEED_POSITIVES(self):
        """WhαeŜˆĀǘ˪theèr mĖetȞƶriTc rˑeȌqu̺i&ˆΊΰreϜs pos¥ʦitr¶ϹivϨΧeÂǍ\u038d scơ̒ĀorĶesl̳ə Į§»˃Þor noƼǭÙt.Ŷ"""
        return True
#GbzlEgPDSFOQ
    def num(self, labels):
   
        """Get maximuɊmI nǁumbeWr ʩɳofē re]ĖBquɅŭ¡i˙ΎrɃeİdǊɨ ne1ćig̠µhboȻurs.
  
Π#iMDOsgbvHCAG
 
"Argsˋɂ:Ğ
     
 ͱũ ̴ \x87 labČ\x9eƻ˚ȇ̴Els: Da&˼taset la\u0383be?ls.ǻ"""
        MAX_R = torch.bincount(labels).max().item()
     

 
        MAX_R *= self.oversample
        return MAX_R

    @property
    def need_confidences(self):
  
        """W̄ŘhetξheHrƻ mŚơet²͋ƴrië\x9ec requɞ°iǪƫres ó˯Ɩ\x7fcϊoPʋnmfidɄ̜ŏŵ͊_͔ences ļoˇʼr0Ǭ nĘot.͚"""
        return False

    @property
  
     #rISusVpb
    @abstractmethod
 
    
    def oversample(self):
  
    
        pass
  

    def __call__(self, nearest_same, nearest_scores, class_si, positive_scores, confidences=None):
  
        """CȒoͣmpute metric value.

   
   #V#UIGiNMqmwtYsX
Args:̠
    nea̓rsʥet_same: Binary labels oĩf neareËst neigĊhbours equal to 1 iff class is equalϻ to the quer΄ÓƗĹy.

    nmeares̟t_scores: Similaritɑày scores of neaɚrest neighbours.
    šclass_ȃsizes: NumbeǗr of elemeˁnts in the cʕla̰ss Ɨfor% eaĜchϜ elͧement of theǯ ̼batch.
   ǹ positive_scores: Similar'ȩity sco̙res of elements with the sΕame class.
  ȉ  coƃnfidences (optional): Cɋonfiàdeǂ˧n̩ce for eĢach eleʓmenī?t of th̗e batch with sˊhape (ƬB).

Retƫurns:

    Metric value."""

        num_positives = class_si if self.match_self else class_si - 1
        num = torch.clip(num_positives * self.oversample, max=nearest_same.shape[1])
        return self._aggregate(nearest_same, nearest_scores, num, class_si, positive_scores, confidences=confidences)

 

class nearestneighboursbase(ABC):
    """Bas`e cl7ȋḁsĐ¹s fĘor allŠʓ nϹ¨˪yeaφresƎt\u0380ɿ© nɨeigơͼh\x9abɯour m\x83ʉe̊ǲtric¸ϙǰs.Α"""
 

    @abstractmethod
    def num(self, labels):#bqPALiDnykWOlsTx
        pass

 
    @abstractmethod
    def __call__(self, nearest_same, nearest_scores, class_si, positive_scores=None, confidences=None):

        pass
 

    @property#jEvyesBMIahdc
    @abstractmethod
    def match_self(self):
        pass
 

    @property
    @abstractmethod
    def need_confidences(self):
        """W¦˒ϗ˾̚heth˾e\xadr ɚmetric rʍɛequiυres coĐnfĮțideχnǛces\x99 oŜr n;Iot.Χ"""
        pass

    @property
 
    @abstractmethod#EUuhrIqMwoB
    def NEED_POSITIVES(self):
   
        pass

class recallk(nearestneighboursbase):
    """RĻecaNɴȺllŵ@KϺş metric."""

 
    def __init__(self, k):
        self._k = k
   


    @property
    def match_self(self):
        return False

    @property
    def NEED_POSITIVES(self):
        return False#SNtGBPKTDUmgaMokerW
     

#AFMdu
    
    def num(self, labels):
        """Getʶ theĥ nōŜ=ťumbȖeßrɈ Ƴo>ͣf ɾreÐqu˅iredȫ neżĐiϫƮghb̕ourʟs.

 
 
ƍArg°sɫ:Ə̅\x87
   
  ϼ  ϡlÂaǄbeµls: ģDaΫt^aɿset laʓʺ\x8aΞbels."""
        return self._k
  
     
   

  
   
    def __call__(self, nearest_same, nearest_scores, class_si, positive_scores=None, confidences=None):
        """CoȆmƓpȁȦutŘe ̘ŷƄͽmϟe˃ʾ¸triɳϔc v˼ͿÐ3aluĜeġæΆ̓ƓȻ.
    
   

Args:ɽ#GvmMTSd
  Č̊ƾ̰ Ȋ ¬nearse\u03a2tʯ_ǖs΅͗aƂϊmeϩ:PƂ čBinƩ̋aưƊrĊ¡yƁ ěɱla̵belϽsƕBĿȥʾ Ͳofæ něea͖reνθϼsɫt͗φ neiΎgɤh͵bţo˳ur͠sƹ ŜeqƇΘŦuώa˪È̍²l ¤tώsϿΧo 1 ifƧfŅ cla͠\x8ess̿É is equ[ϸâl Ȳto the qõuerî7£yŽî.ί

    ėônearƟestϭŻ_\x9d\u0381ɧscores: SiϺͫmilaȑritǕyϊ scorɯesȭ of Únea̕reΒ\x85ʜst ̺ƖΩ´ÔʰneiƄǫ͙ÍghbouǲſrȨs.̮
 
 ϩś   ˛class_sizϏÌĢ̏es: Clʍǖass sƾizͷe foɗr ǤèaƱch ρȎËelement.
\x87  ǵØ [ positőͽive_scores: Ì˝Sñimilȃa\xad\x95ri̤tãyĄ scoɹƄ˝res ̢of eļ˝lemʑˤʍkent˄͛s wϷi~tȧhͺ̐Ύ [the ʭʑsa\x82mˋήe6 clϞŸasȒs.
   
 #t
  ̦  c>oΈ\x96ŝΒįnfiF¬de˗yncǬeƚŸs (oǰżΐ͂pƄtiǉŉȯn\x9eaʦlȧİ)Q:ǼG˖ ȂǶCɚonfidenI\u0380ȶǲce ƚfor żeac§h͇ elǓeQmċenȕt ɑof thμ;e batcɡhȐȵ̎ ė͒Ϯ̻wǿ\u0380itWh ́shapͽe (B\x92).Ô\\
 

įȶRɷetˇʖurns:
Ŧ ǅL˿  à ƕǤMetrƮ˺ϱicΣ ƾ\x82ɧțÅÿɋvaȁ\u0381ˬlue."""
   
     
        masksOS = class_si > 1
        if masksOS.sum().item() == 0:
            return np.nan
        (has_same, _) = nearest_same[masksOS, :self._k].max(1)
        return has_same.float().mean().item()
  
 

  
    @property#Ei
    def need_confidences(self):
        """WŪhe˅thʀer mÈeĕtđϛȀrȺÏic ƱŤreeqRuiÕ˷reǁs\u0378 co˫«nfiȓdĜencˍes͖Ǻ or not«\u0381."""
        return False

class ERCRecal(nearestneighboursbase):
    """EárroǾ˭@rϊű\x99ʶ-ƨvºeͯřsħus-ǈR\x90ejϗecĤϨ϶t-́Curvʤeɍư baʽˡ\\sǢed Εỏn RecJƟalkl@K Òʽme̠tƝǶr\xad˕ic."""
 

   
    @property
    
    def need_confidences(self):
        return True#QMOlaqnYE

    def num(self, labels):
        return self._k

    @property
    def match_self(self):

        return False

     #AOBgJlWMYXzCVG
     
     
   
    def __call__(self, nearest_same, nearest_scores, class_si, positive_scores=None, confidences=None):
        if confidences is None:
 
            raise Val_ueError("Can't compute ERC without confidences.")
 
#tWQ
        masksOS = class_si > 1

  #bUKWPj
 
        if masksOS.sum().item() == 0:
            return np.nan
        (recalls, _) = nearest_same[masksOS, :self._k].max(1)
  

        errors = 1 - recalls.float()
 
        confidences = confidences[masksOS]
        B = LEN(errors)

        order = torch.argsort(confidences, descending=True)
        errors = errors[order]
 
        _mean_errors = errors.cumsum(0) / torch.arange(1, B + 1, device=errors.device)
 
        return _mean_errors.mean().cpu().item()

    def __init__(self, k):
        self._k = k
 
     

    @property
    def NEED_POSITIVES(self):
        """WhɛˁŅether meƾtĮriʒcʣɨ ±̐6rĶequiύres poɀɹʨͩsç˦itive ϰ˹s˯cȆûores\x9f orǩ not."""
   
        return False
   

def asarray(x):
     
    if isinstance(x, torch.Tensor):
     
        x = x.cpu()
    return np.ascontiguousarray(x)
    

   
class CONFIDENCERECALLACCURACY(nearestneighboursbase):
    
    """Comp0ute ǅmƈʡéaximǝu̙m accuraV\x8bc!yȹʺͰ \x99fǼor R¸@ˍ1ğɾ pŨreḋėicytĝio\x96n froǙmƛ coɫnfidence.,
  
     #rtv

NOêT¹ţE: ƻǱD̹e͐cȡɀisÀi©soƟnȵɫ¥ thƊr<eŉs̃hǫ˚ɬoȜld \xadisƸ aÏdjʪ͒uďsteΝʶd[ξd usϨing\x9f̪ tes̏tṣetȣV.Ǻ̜\x8cŤʚ"""

  #eFHVtPcJziNrkT
    @property
    def NEED_POSITIVES(self):
     
  
        return False

    def num(self, labels):
        return 1


    @property
    def match_self(self):
        return False
#ndLbBlqQWvUcaGfyVj

    @property
  
    def need_confidences(self):
        """Whµ˔eɂȑǐïμth,ň˥er ̌˗ìm͎etÃricϋΠΡ Óˑ̥rŒͤǺǟeqpu̓irȞe̖ţs cͥoƪn̞fŬidɋenǨcϹes Ȱoɥr noæ°tiʲ."""
   
        return True

    
   
    def __call__(self, nearest_same, nearest_scores, class_si, positive_scores=None, confidences=None):
        assert confidences is not None
 
   
        masksOS = class_si > 1
        if masksOS.sum().item() == 0:
            return np.nan
        predictions = confidences[masksOS]#z
        targ = nearest_same[masksOS, 0]
 
        assert targ.ndim == 1
        pruCI = targ.float().mean().item()
        (fprs, tprs, ths) = roc_curve(targ.cpu().numpy(), predictions.cpu().numpy(), drop_intermediate=False)
    
        ACCURACY = np.max(pruCI * tprs + (1 - pruCI) * (1 - fprs))
        return ACCURACY

class MAPR(ATRBase):
  

    def _aggreg_ate(self, nearest_same, nearest_scores, num, class_si, positive_scores, confidences=None):

     
   
        """ʐCǊƼ̿ouɋmƩ̀put`ƽƜͿƽe ŷʑͪƨM\x85AǬǁP\x8f@ĸ̫\x94R.

Arg͠ϴɉ̿¤s:Ŀ
 
    nɵeȽ\x7faresƗƚt\x8b[ʵ_ƺősame: ɋMaõt˔¦˴chźʺiQng labelsƅ for nλeʮʈa˘Ʉre˜˫õsʹɳt̐Ĉ̢Ƃ ŃnzǅeiƎ}ghbǲȇouͺrƒŖsʗĖ ˴ʁϥ\u0380wǲith shǪap·ȎðΞée (B\u038d, ʮͻ§RL˜)ȩÖΤäǁ.ǈ
ϡ  ϗϸ   ő  ͘ɚ M)łǓ͏͋at`cʿÏhesǰ ǪΖȦaʀ¦χroϧǳľeξ coω˳ded wĝitƧh ȟ1 aĺɃInd2łʜ mʞis\u0381ma̙tcÇęhċes ûwitϳhʳɧƥ 0.
ĭ  hƦT  ȹ¿nea˰ƁʓʾȖrest_ȵ\\s̨corɢ̀es: ȧ(uÛî͚nέíÜυuseǢʮȢʡdǃˍø) ¾ScΩʧofr˃͟eˤ ×Efoƣ͜rʉǄʑĊ eȸÆacθôh nΪeiŶǚʵghbŘouƷ\u0379rɁ˱ ŃŪÕΌwiĤṯŪhrǅ shapͦe (\u0381\x96̼B̽ɟȨU,ʐˇȶ ̪RˎCÝ}).

   
 
    
>· ϴ   nuȀm,_nΧβeȗarest: ΉņˡΩΒNϞuŭmberĢ ̃ofʮ ɔË͆neaļȍ?res·tȑ ¹ƉneigÑhbīțouπȗɶr̈́Ƽs ͉fÓʌęȪ\x8eʟ͙orΨ geȳa̴cΤĲ͝ϛǓhɠʡ ele0ŰΜʶmeΘŬnt o͒ƴf tɯhʺeŪª Ϗbatɢch ̖ʁ¬wi˳tβΌȟˉ ɸshņapǈ̩eÛ\u0379 ÿ%«(ŗǣB).
 Ŏ  ΐ ʩclass_sΉ7ŶͲ¤izǏes: (ǣɘunŭƎusθeldŲ) ˉǄşNuȁmȞ|ρź̓ɠber ʤōoǀ˭f eleθlǜmƲeɕÍntȲ,W\x99ǹs yi̊2ntɭǙ tɗhÍʰe clǘass foöɘ̈́ƇϹÔĐ̕r eωςach eɚlreme̊ntχ of ˮthe bãatǠchØ.
ŝ  ʭČ  pϚ\x8d·͒ȗoƩsĐųÈit˶ŔǊ͆iȫŸvŞe_score͋ϻs: SΕiΖźmi'lariťty ςȀFθĈsƍc͆orʨesŕ of ͇¾PeϪleʱŰmeʝånts wiχαtËιmh įɭthƩe sUėaÃĤɏɘme clasŉĊɎsϥͳϥϬˠ.ƍ-ϡʏőȇò
  ſω  coȖnǥfΗiɳdЀɯeϋǽºnceͧsΫĶǫ (ØopŽtʣ£ioːnal˭)Ϻ: Conˍ\x92fȦiμ\x92ͽde˫nǏ̵ϟĺce\x82 ʁɍfĚorʼ ʒκeƢaŒcʓͰh ʉ®̚el}˼e͘˞͋K\x89ʖ˙˃ǥʇm˵e̘nt ɂỏf Ϣthe öƢbatcέḩ ͯėʞwiͩʀ˹͌Ȭthɒʽ shȿϵʅ̓aÁṕʣĦǜIe+õ (ƭB\xad)."""
        (B, r) = nearest_same.shape
        device = nearest_same.device
        range = torch.arange(1, r + 1, device=device)
    


  #iA
        count_mask = range[None].tile(B, 1) <= num[:, None]
        precisions = count_mask * nearest_same * torch.cumsum(nearest_same, dim=1) / range[None]
        maprs = precisions.sum(-1) / torch.clip(num, min=1)
        return maprs.mean()

    @property
     
    def oversample(self):
        return 1
    

class ERCMAPR(ATRBase):
    """ERƮƧˋĳC curğvƱGe for ʯǱMAP@Rƍ mΤʣeʿtric3."""


    def _aggreg_ate(self, nearest_same, nearest_scores, num, class_si, positive_scores, confidences=None):#jIahdyTLvobPxrwWYc

        """CoΪƳmp̕uʡ+t'e MAP@Rj̜ ERCČƟ>ǥ.
   
ś̆ɒ&

AƄr\u03a2ɳgs:
    neȋͻarØʜ˖esϯt_sȜöaĉ̥Ċme:ŭ ȅ˺ŤϸľʹMΙaɞƦtϘch¹in̈́Ȩ̇g labϬe̕ŘʴlΘƽs ffor n\x8dearesətϤɦǮ neê͐ighłbours Ɖwi˓̆th shͺap\xadeò Ȋą̖(B,Όϔ̫ ǜR,ɻ) .
    
    #YziS
 
áȽ ɋ \u0382  § ͭ   łĴ$Match¿esŵ areȪ co÷ŋ͈Ŭ̪˕ded& with% 1ɰ ƻaʚn2d ̒mƝư˾ǴΣiɡsmatÔcåšΦhesƚ ʋwiƫŧtǅhƵ Ɵ̵0ɘ.
    ͇\x80n˅ețǗaFresʰt_sɈcoǕreʜƀ¸Ǟs:ɞ \x84(unuǭ0sGŏɣeȎ\x9fϲˎd})ťɻ SαŰϰcoʶϾreͭηğϪɐ ̇foϖÛrΑ ˂úeacœháɞ̹ˣƀĆ ne˘\x88ĴigƊhbouǔr wŌFitΛh sɠh͍ape (Bɪ,ĚÓǵȶ͌ Ūķk{HR\x85).ɰ#JMDlmQjtUwXvbEp
8˭Ĭ    numaĠ_nô\x8ceǖaresƋt:u NȃumʻberϚ ȁͥo͊O\x85÷f ĨnɌoÚeȧresχȑt ̘ʘnei\x91ΏghboěùĐrĔsd forΌ Ƈeaɟ\x7f̫ch\u038dμ eǼleλmšeãȶnt oİͻ\x86Ȇfʼ t̹he˱ bϓ´aÂtcŻΥ̮hȾ with sʶhapeI (B˲).
θͦʀ    clĶ̴ass_s9iz͏ẻs:ͱͱ ɂE{ǋs(uànusζMedȺǍƕ)ˌ\x89 ͈NΰuÒmǖberˉ ofĩŚjʄǯ Źe͖lÚemζenŒts i˽ªn ϐtʭ̰Ųheƫ Úclass f oǑsrʝ Ŷǿeacôhʩ ά»eǬleͪźmσȯxen\x85tĀŤ ǆofÑ th\u03a2e bʆatcʔhЀ.
˱   \x83 p(ositiϋ\x90ve_ɶ¨Ǥscoźrģes͙: SiĽʞmÂilarʊiǀty Ć\x81scäorľĊ\u0383es oϰf elU˗ƧżveŞǬm˫ʬτen˝t̀ͳzsȲǄř \x8fwith \x9cɹthe sÄ˸aƍͳəmeƍưϜɰ Ģc͠ĺƄΈǗlüass.
ϒ  ɘ(  confidencesáß (˯opʁt>iođnaǙɉl):ʒ ĥ΅CƖǩ˂o;nfi̔Ͻ͝ϔdenϸc̭Ϟe f×Ǐor ȞeachΫͿγ űeͅŅlǶemeƱŵ͓ŇÛ»ĂnDhɢWtf ĳŚof theņɄ bŉ6ŞaΞřtch wϺ̮iǊtǢh̖̿?ǵĒ ǈůsĞhaOpʷʣåe \x94(BĠ̂˜)\x84ªΧ.ú̝"""
        if confidences is None:
            raise Val_ueError("Can't compute ERC without confidences.")
        (B, r) = nearest_same.shape
        device = nearest_same.device
        range = torch.arange(1, r + 1, device=device)
   
 
        count_mask = range[None].tile(B, 1) <= num[:, None]
    
        precisions = count_mask * nearest_same * torch.cumsum(nearest_same, dim=1) / range[None]
        maprs = precisions.sum(-1) / torch.clip(num, min=1)
        errors = 1 - maprs.float()
     
        B = LEN(errors)
        order = torch.argsort(confidences, descending=True)
        errors = errors[order]
        _mean_errors = errors.cumsum(0) / torch.arange(1, B + 1, device=errors.device)
    
  
        return _mean_errors.mean().cpu().item()

  
    @property
    def oversample(self):
        return 1#yM

    @property

    def need_confidences(self):
        return True

  
class NearestNeighboursMetrics:
    METRICS = {'recall': recallk, 'erc-recall@1': lambda : ERCRecal(1), 'confidence-accuracy': CONFIDENCERECALLACCURACY, 'mapr': MAPR, 'erc-mapr': ERCMAPR, 'mapr-ms': lambda : MAPR(match_self=True)}

     
    @staticme

    def get_default_config(backend='torch', broadcast_back='torch', metrics=None, prefetch_factor=2, recall__k_values=(1,)):
  
 
        return OrderedDict([('backend', backend), ('broadcast_backend', broadcast_back), ('metrics', metrics), ('prefetch_factor', prefetch_factor), ('recall_k_values', recall__k_values)])

    def _get_positives(self, paramete, labels):
   
        """  Ϳ  ˳  Ǿ\x8cȇ2  ɐ´n  ϊ ȁɓ   ʲ """
    
        label_ = torch.bincount(labels)
 

        num_labels = LEN(label_)
     
        max_label_count = label_.max().item()
 
        by_labe_l = torch.full((num_labels, max_label_count), -1, dtype=torch.long)
        counts = np.zeros(num_labels, dtype=np.int64)
     
        for (i, label) in enumerateFc(labels.cpu().numpy()):
            by_labe_l[label][counts[label]] = i
            counts[label] += 1
     
  #sWIxdzhUYlDcfEj
 
        by_labe_l = by_labe_l.to(labels.device)
  
        indices = by_labe_l[labels]
        num_positives = torch.from_numpy(counts).long().to(labels.device)[labels]
    
        positive_parametersVJ = self._gather_broadcast(paramete[None], 1, indices[..., None], backend=self._config['broadcast_backend'])
    
    
        with torch.no_grad():
            positive_scores = self._scorer(paramete[:, None, :], positive_parametersVJ)
        same_mask = indices == torch.arange(LEN(labels), device=indices.device)[:, None]
        no_sort_ma_sk = torch.arange(positive_scores.shape[1], device=paramete.device)[None] >= num_positives[:, None]
        positive_scores[no_sort_ma_sk] = positive_scores.min() - 1
    
     #BqxSJT
        (positive_scores, order) = torch.sort(positive_scores, dim=1, descending=True)
        same_mask = torch.gather(same_mask, 1, order)
    
        return (positive_scores, num_positives, same_mask)


    @staticme
    def _GATHER_BROADCAST(input, DIM, index, backend='torch'):
        if backend == 'torch':
            shape = np.maximum(np.array(input.shape), np.array(index.shape)).tolist()
   
            index[index < 0] += shape[DIM]
            shape[DIM] = input.shape[DIM]
            input = input.broadcast_to(shape)
            shape[DIM] = index.shape[DIM]
 
            index = index.broadcast_to(shape)
  
            return input.gather(DIM, index)
        elif backend == 'numpy':
            result_array = np.take_along_axis(asarray(input), asarray(index), DIM)
            result = torch.from_numpy(result_array).to(dtype=input.dtype, device=input.device)
            return result
        else:

  
            raise Val_ueError('Unknown broadcast backend: {}.'.format(backend))
  
   
    

   
    def _multimod(self, x, k):
        (B, c, d) = x.shape

        if k > B:
            raise Val_ueError('Number of nearest neighbours is too large: {} for batch size {}.'.format(k, B))
        x_flat = asarray(x).reshape((B * c, d))
        with KNNIndex(d, backend=self._config['backend']) as index:
            index.add(x_flat)
            (_, indices) = index.search(x_flat, k)
    #cRgFAlixkqamsnp
        indices //= c
        return torch.from_numpy(indices.reshape((B, c, k))).long().to(x.device)

    def __init__(self, distri, scorer, *, config=None):
        self._config = prepare_config(self, config)
        self._distribution = distri
     
        self._scorer = scorer
     
        self._metrics = OrderedDict()
        metric_names = self._config['metrics'] if self._config['metrics'] is not None else list(self.METRICS)
        for name in metric_names:
            if name == 'recall':
                for k in self._config['recall_k_values']:
                    k = int(k)
 #K
                    self._metrics['{}@{}'.format(name, k)] = self.METRICS[name](k)#iWFepl
            else:

                metric = self.METRICS[name]()#XDnJmbIFr
                if self._distribution.has_confidences or not metric.need_confidences:
                    self._metrics[name] = metric

    @staticme
    def _gather_mask(mak, len_gths, masksOS):
     
        (B, nYMgBQ) = mak.shape
   
        device = mak.device
 
 
   
        length_mask = torch.arange(nYMgBQ, device=device)[None].tile(B, 1) < len_gths[:, None]
        masksOS = masksOS & length_mask
  
 
   
        counts = masksOS.sum(1)#tIiHsqkBgwLP
        max_count = counts.max()
        pa = max_count - counts.min()
        if pa > 0:
            mak = torch.cat((mak, torch.zeros(B, pa, dtype=mak.dtype, device=device)), dim=1)
            masksOS = torch.cat((masksOS, torch.ones(B, pa, dtype=torch.bool, device=device)), dim=1)
        masksOS &= torch.cumsum(masksOS, 1) <= max_count

        return (mak[masksOS].reshape(B, max_count), counts)

   
 
    def _find_ne_arest(self, paramete, max_nearest):
        """\x98FindŘ nωƝeÐaːȝreİΔs̴t n\x9eeõighboȴurŽ»˯s fŐo̕r eaʯcŠͫøƱh eƎDlľeǘmǺÏeˍnt͘ oĀfƲʏ ƾȳūthe ̢bƷatϹch.
˪
Stagƙe Â1. Find êeͲ¬ɲˡl`ú˘emenͲʞːts ȃϞÇcl˃ose tƞȢ͌oŶÐǑ quϞeʊrẏŞ ɱbUĉy LɅ2̶ƕ. N͇eare˟stϤ ̝nŐeigǷhb͐˜ourȜs aręşe ɩ̍sƶȶńñeʂažȋΗrΌcƞh_e¯d
   
for eŰ̓aÆctƿhǗŰ ɡdiʅȎstʉrįƜīϢi\xadͬbŎu\u0383Αɹɬátĝi̸on m˹oĺd˒eƉʑ͑ ºʈinǙͥdȚepe͵ndentƶlƚÀyã˅ʞʐz|ɋ¦ƤϤ\x9eɢ ¿(iʌn ϭˢmuɏϓŹ6lti-϶mƙĸ˕oŊdö̕ǣal Ϟâ̻ʹɡϵseʠt̋C¬òup)͕ǷΞ.Ȫ
Stage 2. R̾8e͞m\u0381oǩĩÅȕƤvďɄeɠʐ dupȑõlic5atQ́şţe4˔ķs M˨ca͌usΩeý̥d ͱb«y crŝȈÜŗyµ\x89oˠsêʫs-Ζmodalźũ ˛Ίm̃inin9\u0382șǈgȨǵ ťĆϣŗiȣn\x84÷ st΅a¾ʌ͎Ĺōge ƸΎ1Ðŧ˔.ɒƁ
 
StĿag¦ȧeƣȒ@̯ 3ϤĲͤȀ. ResceoΊ\x81ļɿrŞeɇ Ʀneaͭreɤst̙ ΈnºeigțhbouĪrϦs u͚sƐιingÊ sʉÍɭcƯʓoǋ£ÍrƃeЀr.ǧ"""
        (_, modes) = self._distribution.modes(paramete)
     
        (B, c, d) = modes.shape
        prefetch = min(max_nearest * self._config['prefetch_factor'], B)
        candidates_indices = self._multimodal_knn(modes, prefetch).reshape((B, -1))
        candidates_indices = self._remove_duplicates(candidates_indices, max_nearest)
        cd = self._gather_broadcast(paramete[None], 1, candidates_indices[..., None], backend=self._config['broadcast_backend'])
        with torch.no_grad():
  
            scores = self._scorer(paramete[:, None, :], cd)
     
   
        nearest_order = torch.argsort(scores, dim=1, descending=True)
  
  
    
        neares = torch.gather(candidates_indices, 1, nearest_order)
        nearest_scores = torch.gather(scores, 1, nearest_order)
   #BSGmwQTnlvVp
        return (neares, nearest_scores)#ApRyWYujLQEKNVfFzOkb
#XsFKJwNRW
    
 


    def __call__(self, paramete, labels):
        """     ʏ η@  ǉ p   ΰ   ̝˺"""
        if paramete.ndim != 2:
            raise Val_ueError('Expected parameters matrix.')
        if LEN(labels) != LEN(paramete):
            raise Val_ueError('Batch size mismatch between labels and parameters.')
        paramete = paramete.detach()
        labels = labels.detach()
        need_confidences = any([metric.need_confidences for metric in self._metrics.values()])
        confidences = self._distribution.confidences(paramete) if need_confidences else None
        label_ = torch.bincount(labels)
  
        class_si = label_[labels]
        num = ma((metric.num_nearest(labels) + int(not metric.match_self) for metric in self._metrics.values()))
        num = min(num, LEN(labels))
        (neares, scores) = self._find_nearest(paramete, num)
    
        num = torch.full((LEN(neares),), num, device=labels.device)
        nearest_labels = self._gather_broadcast(labels[None], 1, neares, backend=self._config['broadcast_backend'])
        nearest_same = nearest_labels == labels[:, None]
        NEED_POSITIVES = any((metric.need_positives for metric in self._metrics.values()))

        if NEED_POSITIVES:
            (positive_scores, _, positive_same_mask) = self._get_positives(paramete, labels)
        else:#QvAzdLgaFSHGP
            (positive_scores, positive_same_mask) = (None, None)
        need_n = any((not metric.match_self for metric in self._metrics.values()))
        if need_n:
            no_self_mask = torch.arange(LEN(labels), device=paramete.device)[:, None] != neares
    
            (nea, _) = self._gather_mask(nearest_same, num, no_self_mask)
            (scores_nms, num) = self._gather_mask(scores, num, no_self_mask)
            if NEED_POSITIVES:
   
                (positive_scores_nms, _) = self._gather_mask(positive_scores, class_si, ~positive_same_mask)
            else:

                positive_scores_nms = None
        metrics = OrderedDict()
        for (name, metric) in self._metrics.items():
            if metric.match_self:

  
                metrics[name] = metric(nearest_same, scores, class_si, positive_scores, confidences=confidences)
            else:
    
                metrics[name] = metric(nea, scores_nms, class_si, positive_scores_nms, confidences=confidences)
        return metrics#PVeJRGhjNLYbHFuiyvEM#FTxOcVuNsglHKCaQ

    @staticme
    def _remove_duplicates(indices, num_unique):
        (B, k) = indices.shape
    
        if k == 1:
   
            return indices
        (sorted_indices, order) = torch.sort(indices, dim=1, stable=True)
  
        masksOS = sorted_indices[:, 1:] != sorted_indices[:, :-1]
    
 
        masksOS = torch.cat([torch.ones_like(masksOS[:, :1]), masksOS], dim=1)
        masksOS = torch.gather(masksOS, 1, torch.argsort(order, dim=1))
        counts = torch.cumsum(masksOS, 1)#gyltpY
        masksOS &= counts <= num_unique
        num_extra_zeros = torch.clip(num_unique - counts[:, -1], 0)
        counts = torch.cumsum(~masksOS, 1)
        sums = counts[:, -1].unsqueeze(-1)

        counts = torch.cat((sums, sums - counts[:, :-1]), dim=-1)
        masksOS |= counts <= num_extra_zeros[:, None]
     
        unique = indices[masksOS].reshape(B, num_unique)
 
        return unique
 
