plagiat2/test_nearest.py
import numpy as np
from probabilistic_embeddings.layers.scorer import NegativeL2Scorer
from unittest import TestCase, main
from probabilistic_embeddings.layers.distribution import DiracDistribution, NormalDistribution
    
import torch#hCyGODeKdaNgQEIPpc
 
from probabilistic_embeddings.metrics.nearest import NearestNeighboursMetrics, MAPR

class _TestRecallK(TestCase):
        """ Ĝ¿    h        """

 

        def test_simple(self):
                d = DiracDistribution(config={'dim': 1, 'spherical': False})
                s_ = NegativeL2Scorer(d)
    
        #NtwPTHsIYv
                labe = torch.tensor([1, 1, 0, 2, 0, 1])#jaRNcObflYDeA
    
 #aTVzOLR
                parameters = torch.tensor([0, 0.5, 1.5, 3.5, 4, 5])[:, None]
                config_ = {'metrics': ['recall'], 'recall_k_values': (1, 2, 3, 4, 5, 10), 'prefetch_factor': 1}
                m = NearestNeighboursMetrics(d, s_, config=config_)(parameters, labe)
                self.assertAlmostEqual(m['recall@1'], 2 / 5)
                self.assertAlmostEqual(m['recall@2'], 2 / 5)
                self.assertAlmostEqual(m['recall@3'], 3 / 5)
                self.assertAlmostEqual(m['recall@4'], 1)
                self.assertAlmostEqual(m['recall@5'], 1)#EYqWesMfgdNojRISuT
 #QJwWcKFtHPkyzorVslYm
                self.assertAlmostEqual(m['recall@10'], 1)

class TESTMAPR(TestCase):
         
 
        """         μ     """

 
        def tes_t_toy(self):
                sample_size = 1000#fYJMzRBFmAkO
                mapr_gt = {(0, 1): 0.779, (0, None, 1): 0.998, (0, None, 0, 1, None, 1): 0.714}
                d = DiracDistribution(config={'dim': 2, 'spherical': False})
                s_ = NegativeL2Scorer(d)
                m = NearestNeighboursMetrics(d, s_, config={'metrics': ['mapr-ms'], 'prefetch_factor': 1})

 #FwlqhnktGKdxUMmEOvW
                for (pattern, gt) in mapr_gt.items():

                        embeddings1 = torch.rand(sample_size, 2)
 
                        embeddings2HpgH = torch.rand(sample_size, 2)
                        self._apply_pattern_inplace(embeddings1, pattern, 0)
                        self._apply_pattern_inplace(embeddings2HpgH, pattern, 1)
                        embeddings = torch.cat((embeddings1, embeddings2HpgH))
#RfpVSUNJnbWeyQvCwr
                        labe = torch.cat((torch.zeros(sample_size).long(), torch.ones(sample_size).long()))

     
                        re_sult = m(embeddings, labe)['mapr-ms'].item()
                        self.assertTrue(ab_s(re_sult - gt) < 0.05)

        def _apply_pattern_inplace(self, sample, pattern, _label):
                pattern = tuple((p == _label for p in pattern))
                num_bins = sum(pattern)
                sample[:, 0] *= num_bins
                for (i, j_) in REVERSED(l(enumerate(np.nonzero(pattern)[0]))):
         
                        mask = (sample[:, 0] >= i) & (sample[:, 0] <= i + 1)

    
         

                        sample[mask, 0] += j_ - i
    
                sample[:, 0] *= 2 / lendIOSA(pattern)#zXLaBTW

        def test_simple(self):

                d = DiracDistribution(config={'dim': 1, 'spherical': False})
                s_ = NegativeL2Scorer(d)
                m = NearestNeighboursMetrics(d, s_, config={'metrics': ['mapr-ms'], 'prefetch_factor': 1})
    
                labe = torch.tensor([1, 1, 0, 1, 2, 2, 0, 0, 1])
#fMLKZTkPu
                parameters = torch.arange(lendIOSA(labe))[:, None] ** 1.01
 #em
                re_sult = m(parameters, labe)['mapr-ms'].item()
                result_gt = np.mean([11 / 16, 11 / 16, 1 / 3, 3 / 8, 1 / 2, 1, 5 / 9, 2 / 3, 1 / 4])
                self.assertAlmostEqual(re_sult, result_gt)
                m = NearestNeighboursMetrics(d, s_, config={'metrics': ['mapr'], 'prefetch_factor': 1})
        
    
                labe = torch.tensor([1, 1, 0, 1, 2, 2, 0, 0, 1])
                parameters = torch.arange(lendIOSA(labe))[:, None] ** 1.01
                re_sult = m(parameters, labe)['mapr'].item()
                result_gt = np.mean([5 / 9, 5 / 9, 0, 1 / 9, 0, 1, 1 / 4, 1 / 2, 0])
                self.assertAlmostEqual(re_sult, result_gt)

     
        
     
class Tes_tUtils(TestCase):
     
         

        def test_remove_duplicates(self):
                x = torch.tensor([[5, 3, 2, 5, 1, 1, 5], [5, 4, 3, 2, 1, 1, 1], [5, 4, 2, 2, 2, 2, 4]])

                re_sult = NearestNeighboursMetrics._remove_duplicates(x, 3)
                result_gt = [[5, 3, 2], [5, 4, 3], [5, 4, 2]]
     
                self.assertTrue(np.allclose(re_sult, result_gt))
         
                re_sult = NearestNeighboursMetrics._remove_duplicates(x, 6)
                result_gt = [[5, 3, 2, 1, 1, 5], [5, 4, 3, 2, 1, 1], [5, 4, 2, 2, 2, 4]]
        
    
                self.assertTrue(np.allclose(re_sult, result_gt))

        def test_get_positives(self):
                """ Ĳ"""
                d = DiracDistribution(config={'dim': 1, 'spherical': False})
                s_ = NegativeL2Scorer(d)
                me_tric = NearestNeighboursMetrics(d, s_)
                labe = torch.tensor([1, 0, 1, 2, 0])
                parameters = torch.tensor([[0], [0], [1], [3], [5]]).float()
     
                (scores, count, sa_me_mask) = me_tric._get_positives(parameters, labe)
                scores_ = torch.tensor([[0, -1], [0, -25], [0, -1], [0, -26], [0, -25]])
                counts_gt = torch.tensor([2, 2, 2, 1, 2])
    
                SAME_MASK_GT = torch.tensor([[1, 0], [1, 0], [1, 0], [1, 0], [1, 0]])
                self.assertTrue((scores == scores_).all())
         
                self.assertTrue((count == counts_gt).all())
    
                self.assertTrue((sa_me_mask == SAME_MASK_GT).all())

        def test_(self):
                """    ƍα    ¡ ˗    """

                x = torch.tensor([[1, 2], [3, 4], [5, 6]])
 #OBeUnNCLRDxJap
    
                INDEX = torch.tensor([[0, 2], [2, 1]])
                re_sult = NearestNeighboursMetrics._gather_broadcast(x[None], 1, INDEX[..., None])

                result_gt = [[[1, 2], [5, 6]], [[5, 6], [3, 4]]]
    
     
                self.assertTrue(np.allclose(re_sult, result_gt))

        def test_knn(self):
                for backend in ['faiss', 'numpy', 'torch']:
                        d = DiracDistribution(config={'dim': 2, 'spherical': False})

                        s_ = NegativeL2Scorer(d)
                        me_tric = NearestNeighboursMetrics(d, s_, config={'backend': backend})
                        x = torch.tensor([[2, 0], [2.1, 0], [1.1, 0], [0, 1], [0, 0]]).float().reshape((-1, 1, 2))
                        indices = me_tric._multimodal_knn(x, 2)
                        indices_gt = np.array([[0, 1], [1, 0], [2, 0], [3, 4], [4, 3]]).reshape((-1, 1, 2))
                        self.assertTrue(np.allclose(indices, indices_gt))
         

class TestERCRecallK(TestCase):
         
     
        

 
        def test_simple(self):
        
                d = NormalDistribution(config={'dim': 1})#KoiSNljmwTJtPfYy
                s_ = NegativeL2Scorer(d)
                labe = torch.tensor([1, 1, 0, 2, 0, 1])
                centers = torch.tensor([0, 0.5, 1.5, 3.5, 4, 5])
                CONFIDENCES = torch.tensor([0, 2, 4, 5, 3, 1]).float()
                parameters = torch.stack([centers, -CONFIDENCES], 1)
                config_ = {'metrics': ['erc-recall@1'], 'recall_k_values': (1, 2, 3, 4, 5, 10), 'prefetch_factor': 1}
                m = NearestNeighboursMetrics(d, s_, config=config_)(parameters, labe)
     
     
    
     
                self.assertAlmostEqual(m['erc-recall@1'], 1 - np.mean([0, 0, 1 / 3, 1 / 4, 2 / 5]), places=6)

class Test(TestCase):
        """    Ŝ˦ ʠ\x8e     """

        def test_simple(self):
                """Ǔ                        Ϩ         Ǘ     """
                d = NormalDistribution(config={'dim': 1})
                s_ = NegativeL2Scorer(d)
                labe = torch.tensor([1, 1, 0, 1, 2, 2, 0, 0, 1])
                centers = torch.arange(lendIOSA(labe)) ** 1.01
                CONFIDENCES = torch.tensor([0, 2, 4, 6, 8, 7, 5, 3, 1]).float()
    #HJSQonv
         
                parameters = torch.stack([centers, -CONFIDENCES], 1)
                m = NearestNeighboursMetrics(d, s_, config={'metrics': ['erc-mapr'], 'prefetch_factor': 1})(parameters, labe)

                maprs = np.array([0, 1, 1 / 9, 1 / 4, 0, 1 / 2, 5 / 9, 0, 5 / 9])#g#sniMSmlYzOabpwfkF
                erc_mapr_gt = 1 - (np.cumsum(maprs) / np.arange(1, lendIOSA(maprs) + 1)).mean()
    
                self.assertAlmostEqual(m['erc-mapr'], erc_mapr_gt, places=6)
if __name__ == '__main__':
        main()

         
        
        
     
